router 정리
