+++ app.js
@@ -4,8 +4,14 @@ var path = require('path');
 var cookieParser = require('cookie-parser');
 var logger = require('morgan');
 
+// var connection = require('./routes/db');
+
+
 var indexRouter = require('./routes/index');
+var loginRouter = require('./routes/login')
 var usersRouter = require('./routes/users');
+var reservRouter = require('./routes/reservation');
+var registerRouter = require('./routes/register');
 
 var app = express();
 
@@ -21,6 +27,9 @@ app.use(express.static(path.join(__dirname, 'public')));
 
 app.use('/', indexRouter);
 app.use('/users', usersRouter);
+app.use('/login',loginRouter);
+app.use('/reservation',reservRouter);
+app.use('/register',registerRouter);
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
