+++ views/chanwoong/login.ejs
@@ -22,7 +22,7 @@
                         <div class="row">
                             <div class="col-md-9 col-lg-8 mx-auto">
                                 <h3 class="login-heading mb-4">Welcome back!</h3>
-                                <form method="post" action="/main" >
+                                <form method="post" action="/login" >
                                     <div class="form-label-group">
                                         <input type="email" id="inputEmail" name="userid" class="form-control" placeholder="Email address" required autofocus>
                                         <label for="inputEmail">Email address</label>
