+++ views/chanwoong/register.ejs
@@ -41,6 +41,7 @@
         <div class="row probootstrap-gutter40">
             <div class="col-md-8">
                 <h2 class="mt0">Register Form</h2>
+                
                 <form action="/register" method="post" class="probootstrap-form">
                     <div class="row">
                         <div class="col-sm-6 form-group">
