+++ views/chanwoong/reservation.ejs
@@ -44,7 +44,7 @@
             <hr style="width: 100%; background-color: #903479; height: 2px;">
 
 
-            <form action="/search_room" method="post" class="probootstrap-form ">
+            <form action="/reservation" method="post" class="probootstrap-form ">
 
                 <div class="row">
                     <div class="col-md-2">
